refactor: rename `meme` to `emoji`
