+++ src/components/modal/PublishModal.js
@@ -230,7 +230,7 @@ class PublishModal extends Component {
       Keyboard.dismiss();
     }
 
-    if (selectedPanel === 'meme') {
+    if (selectedPanel === 'emoji') {
       this.setState({
         keyboardAccessoryToBottom: 0,
         selectedPanel: 'keyboard'
@@ -336,7 +336,7 @@ class PublishModal extends Component {
               removeImage={imageIndex => this.removeImage(imageIndex)} />
           </View>
         </KeyboardAwareScrollView>
-        {(this.state.isContentFocused || this.state.selectedPanel === 'meme') &&
+        {(this.state.isContentFocused || this.state.selectedPanel === 'emoji') &&
           <KeyboardAccessory
             style={{ bottom: this.state.keyboardAccessoryToBottom }}
             selectedPanel={this.state.selectedPanel}
