+++ src/constants/emojis.js
